feat: move DemoPage component to a separate file
